requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
discord.py==2.3.2
dnspython==2.4.2
exceptiongroup==1.2.1
executing==2.0.1
frozenlist==1.4.0
google-ai-generativelanguage==0.6.4
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-generativeai==0.5.4
googleapis-common-protos==1.61.0
googletrans==4.0.0rc1
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipython==8.24.0
jedi==0.19.1
libretranslatepy==2.1.1
lxml==5.2.1
matplotlib-inline==0.1.7
multidict==6.0.4
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.1
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.1
python-dotenv==1.0.0
pytz==2023.3.post1
pytzdata==2020.1
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tqdm==4.66.4
traitlets==5.14.3
translate==3.6.1
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.13
yarl==1.9.2 
pyjokes==0.6.0
greenlet==1.1.2
sqlalchemy==1.4.31
pillow==10.3.0



